Refactor: split up EntityTypeV3Mapper
