+++ molgenis-api-metadata/src/main/java/org/molgenis/api/metadata/v3/EntityTypeResponseMapper.java
@@ -1,0 +1,151 @@
+package org.molgenis.api.metadata.v3;
+
+import static java.util.Objects.requireNonNull;
+import static org.molgenis.api.data.v3.EntityController.API_ENTITY_PATH;
+import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequestUri;
+
+import com.google.common.collect.ImmutableMap;
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+import org.molgenis.api.PageUtils;
+import org.molgenis.api.metadata.v3.model.AttributesResponse;
+import org.molgenis.api.metadata.v3.model.EntityTypeResponse;
+import org.molgenis.api.metadata.v3.model.EntityTypeResponseData;
+import org.molgenis.api.metadata.v3.model.EntityTypesResponse;
+import org.molgenis.api.metadata.v3.model.I18nValue;
+import org.molgenis.api.metadata.v3.model.PackageResponse;
+import org.molgenis.api.model.response.LinksResponse;
+import org.molgenis.api.support.LinksUtils;
+import org.molgenis.data.meta.model.EntityType;
+import org.molgenis.data.meta.model.EntityTypeMetadata;
+import org.molgenis.data.meta.model.Package;
+import org.molgenis.data.meta.model.PackageMetadata;
+import org.springframework.context.i18n.LocaleContextHolder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.util.UriComponentsBuilder;
+
+@Component
+class EntityTypeResponseMapper {
+
+  private static final String ATTRIBUTES = "attributes";
+  private final AttributeV3Mapper attributeV3Mapper;
+
+  EntityTypeResponseMapper(AttributeV3Mapper attributeV3Mapper) {
+    this.attributeV3Mapper = requireNonNull(attributeV3Mapper);
+  }
+
+  EntityTypesResponse toEntityTypesResponse(
+      EntityTypes entityTypes, int pageSize, int pageNumber, int totalElements) {
+
+    List<EntityTypeResponse> results = new ArrayList<>();
+    for (EntityType entityType : entityTypes.getEntityTypes()) {
+      results.add(mapInternal(entityType, false, true, false, false));
+    }
+
+    return EntityTypesResponse.create(
+        LinksUtils.createLinksResponse(pageNumber, pageSize, totalElements),
+        results,
+        PageUtils.getPageResponse(
+            entityTypes.getEntityTypes().size(), pageNumber * pageSize, totalElements, pageSize));
+  }
+
+  EntityTypeResponse toEntityTypeResponse(
+      EntityType entityType, boolean flattenAttrs, boolean i18n) {
+    return mapInternal(entityType, flattenAttrs, true, true, i18n);
+  }
+
+  private EntityTypeResponse mapInternal(
+      EntityType entityType,
+      boolean flattenAttrs,
+      boolean includeData,
+      boolean expandAttrs,
+      boolean i18n) {
+    EntityTypeResponse.Builder entityTypeResponseBuilder = EntityTypeResponse.builder();
+    entityTypeResponseBuilder.setLinks(
+        LinksResponse.create(null, createEntityTypeResponseUri(entityType), null));
+
+    if (includeData) {
+      EntityTypeResponseData.Builder builder = EntityTypeResponseData.builder();
+      builder.setId(entityType.getId());
+      Package pack = entityType.getPackage();
+      if (pack != null) {
+        builder.setPackage(
+            PackageResponse.builder()
+                .setLinks(LinksResponse.create(null, createPackageResponseUri(pack), null))
+                .build());
+      }
+      builder.setLabel(entityType.getLabel(LocaleContextHolder.getLocale().getLanguage()));
+      builder.setDescription(
+          entityType.getDescription(LocaleContextHolder.getLocale().getLanguage()));
+      if (i18n) {
+        builder.setLabelI18n(getI18nEntityTypeLabel(entityType));
+        getI18nEntityTypeDesc(entityType).ifPresent(builder::setDescriptionI18n);
+      }
+      AttributesResponse.Builder attributesResponseBuilder =
+          AttributesResponse.builder()
+              .setLinks(LinksResponse.create(null, createAttributesResponseUri(entityType), null));
+      if (expandAttrs) {
+        attributesResponseBuilder.setItems(
+            flattenAttrs
+                ? attributeV3Mapper.mapInternal(entityType.getAllAttributes(), i18n)
+                : attributeV3Mapper.mapInternal(entityType.getOwnAllAttributes(), i18n));
+      }
+      builder.setAttributes(attributesResponseBuilder.build());
+      builder.setAbstract(entityType.isAbstract());
+      EntityType parent = entityType.getExtends();
+      builder.setExtends(parent != null ? mapInternal(parent, false, false, false, i18n) : null);
+      builder.setIndexingDepth(entityType.getIndexingDepth());
+      entityTypeResponseBuilder.setData(builder.build());
+    }
+
+    return entityTypeResponseBuilder.build();
+  }
+
+  private URI createEntityTypeResponseUri(EntityType entityType) {
+    UriComponentsBuilder uriComponentsBuilder =
+        fromCurrentRequestUri()
+            .replacePath(null)
+            .path(MetadataApiController.API_META_PATH)
+            .pathSegment(entityType.getId());
+    return uriComponentsBuilder.build().toUri();
+  }
+
+  private URI createPackageResponseUri(Package aPackage) {
+    UriComponentsBuilder uriComponentsBuilder =
+        fromCurrentRequestUri()
+            .replacePath(null)
+            .path(API_ENTITY_PATH)
+            .pathSegment(PackageMetadata.PACKAGE)
+            .pathSegment(aPackage.getId());
+    return uriComponentsBuilder.build().toUri();
+  }
+
+  private URI createAttributesResponseUri(EntityType entityType) {
+    UriComponentsBuilder uriComponentsBuilder =
+        fromCurrentRequestUri()
+            .replacePath(null)
+            .path(MetadataApiController.API_META_PATH)
+            .pathSegment(entityType.getId())
+            .pathSegment(ATTRIBUTES);
+    return uriComponentsBuilder.build().toUri();
+  }
+
+  private I18nValue getI18nEntityTypeLabel(EntityType entityType) {
+    String defaultValue = entityType.getLabel();
+    ImmutableMap<String, String> translations =
+        MetadataUtils.getI18n(entityType, EntityTypeMetadata.LABEL);
+    return I18nValue.create(defaultValue, translations);
+  }
+
+  private Optional<I18nValue> getI18nEntityTypeDesc(EntityType entityType) {
+    String defaultValue = entityType.getDescription();
+    if (defaultValue == null) {
+      return Optional.empty();
+    }
+    ImmutableMap<String, String> translations =
+        MetadataUtils.getI18n(entityType, EntityTypeMetadata.DESCRIPTION);
+    return Optional.of(I18nValue.create(defaultValue, translations));
+  }
+}
